# docs/backend/openapi.yaml
openapi: 3.1.0
info:
  title: Vaeloom Backend
  version: 0.2.0
paths:
...
  /api/v1/workspaces:
    get:
      tags:
        - workspaces
      summary: List Workspaces
      operationId: list_workspaces_api_v1_workspaces_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/WorkspaceResponse'
                type: array
                title: Response List Workspaces Api V1 Workspaces Get
    post:
      tags:
        - workspaces
      summary: Create Workspace
      operationId: create_workspace_api_v1_workspaces_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWorkspaceRequest'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkspaceResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/memories:
    get:
      tags:
        - memory
      summary: List Memories
      operationId: list_memories_api_v1_memories_get
      parameters:
        - name: type
          in: query
          required: false
          schema:
            anyOf:
              - type: string
              - type: 'null'
            title: Type
        - name: domain
          in: query
          required: false
          schema:
            anyOf:
              - type: string
              - type: 'null'
            title: Domain
        - name: status
          in: query
          required: false
          schema:
            anyOf:
              - type: string
              - type: 'null'
            default: active
            title: Status
        - name: workspace_id
          in: query
          required: false
          schema:
            anyOf:
              - type: string
              - type: 'null'
            title: Workspace Id
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
            title: Page
        - name: page_size
          in: query
          required: false
          schema:
            type: integer
            maximum: 100
            minimum: 1
            default: 20
            title: Page Size
        - name: include_superseded
          in: query
          required: false
          schema:
            type: boolean
            default: false
            title: Include Superseded
      requestBody:
        content:
          application/json:
            schema:
              anyOf:
                - type: array
                  items:
                    type: string
                - type: 'null'
              title: Tags
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response List Memories Api V1 Memories Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
        - memory
      summary: Create Memory
      operationId: create_memory_api_v1_memories_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemoryCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemoryResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/memories/search:
    post:
      tags:
        - memory
      summary: Search Memories
      operationId: search_memories_api_v1_memories_search_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemorySearch'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/MemorySearchResult'
                type: array
                title: Response Search Memories Api V1 Memories Search Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/agents:
    post:
      tags:
        - agents
      summary: Create Agent
      operationId: create_agent_api_v1_agents_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
        - agents
      summary: List Agents
      operationId: list_agents_api_v1_agents_get
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
            title: Page
        - name: page_size
          in: query
          required: false
          schema:
            type: integer
            maximum: 100
            minimum: 1
            default: 20
            title: Page Size
        - name: category
          in: query
          required: false
          schema:
            anyOf:
              - type: string
              - type: 'null'
            title: Category
        - name: search
          in: query
          required: false
          schema:
            anyOf:
              - type: string
              - type: 'null'
            title: Search
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response List Agents Api V1 Agents Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/agents/chat:
    post:
      tags:
        - agents
      summary: Chat
      description: 'High-level chat endpoint: auto-classifies intent, routes to the
        right agent,

        runs the agentic loop, and returns the result. If dto.agentName is provided

        (enterprise explicit routing), honors it with high confidence.'
      operationId: chat_api_v1_agents_chat_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatMessage'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/events:
    get:
      tags:
        - events
      summary: List Events
      operationId: list_events_api_v1_events_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/EventResponse'
                type: array
                title: Response List Events Api V1 Events Get
    post:
      tags:
        - events
      summary: Publish Event
      operationId: publish_event_api_v1_events_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventPublish'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/search:
    post:
      tags:
        - search
      summary: Search All
      operationId: search_all_api_v1_search_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/integrations:
    get:
      tags:
        - integrations
      summary: List Integrations
      operationId: list_integrations_api_v1_integrations_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/IntegrationResponse'
                type: array
                title: Response List Integrations Api V1 Integrations Get
    post:
      tags:
        - integrations
      summary: Create Integration
      operationId: create_integration_api_v1_integrations_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IntegrationCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IntegrationResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/documents:
    post:
      tags:
        - documents
      summary: Upload Document
      operationId: upload_document_api_v1_documents_post
      parameters:
        - name: workspace_id
          in: query
          required: false
          schema:
            anyOf:
              - type: string
              - type: 'null'
            title: Workspace Id
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_upload_document_api_v1_documents_post'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
        - documents
      summary: List Documents
      operationId: list_documents_api_v1_documents_get
      parameters:
        - name: workspace_id
          in: query
          required: false
          schema:
            anyOf:
              - type: string
              - type: 'null'
            title: Workspace Id
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
            title: Page
        - name: page_size
          in: query
          required: false
          schema:
            type: integer
            maximum: 100
            minimum: 1
            default: 20
            title: Page Size
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/resumes:
    get:
      tags:
        - resumes
      summary: List Resumes
      operationId: list_resumes_api_v1_resumes_get
      parameters:
        - name: workspace_id
          in: query
          required: false
          schema:
            anyOf:
              - type: string
              - type: 'null'
            title: Workspace Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ResumeResponse'
                title: Response List Resumes Api V1 Resumes Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/workspaces/{workspace_id}/applications:
    get:
      tags:
        - applications
      summary: List Applications
      operationId: list_applications_api_v1_workspaces__workspace_id__applications_get
      parameters:
        - name: workspace_id
          in: path
          required: true
          schema:
            type: string
            title: Workspace Id
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
            title: Page
        - name: page_size
          in: query
          required: false
          schema:
            type: integer
            maximum: 100
            minimum: 1
            default: 20
            title: Page Size
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ApplicationResponse'
                title: Response List Applications Api V1 Workspaces  Workspace Id  Applications
                  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
        - applications
      summary: Create Application
      operationId: create_application_api_v1_workspaces__workspace_id__applications_post
      parameters:
        - name: workspace_id
          in: path
          required: true
          schema:
            type: string
            title: Workspace Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApplicationCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/notifications:
    get:
      tags:
        - notifications
      summary: List Notifications
      operationId: list_notifications_api_v1_notifications_get
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
            title: Page
        - name: page_size
          in: query
          required: false
          schema:
            type: integer
            maximum: 100
            minimum: 1
            default: 20
            title: Page Size
        - name: channel
          in: query
          required: false
          schema:
            anyOf:
              - type: string
              - type: 'null'
            title: Channel
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NotificationResponse'
                title: Response List Notifications Api V1 Notifications Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/connectors:
    post:
      tags:
        - connectors
      summary: Create Connector
      operationId: create_connector_api_v1_connectors_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateConnectorRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
        - connectors
      summary: List Connectors
      operationId: list_connectors_api_v1_connectors_get
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
            title: Page
        - name: page_size
          in: query
          required: false
          schema:
            type: integer
            maximum: 100
            minimum: 1
            default: 20
            title: Page Size
        - name: type
          in: query
          required: false
          schema:
            anyOf:
              - type: string
              - type: 'null'
            title: Type
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ConnectorResponse'
                title: Response List Connectors Api V1 Connectors Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/provider-keys:
    get:
      tags:
        - provider-keys
      summary: List Provider Keys
      operationId: list_provider_keys_api_v1_provider_keys_get
      parameters:
        - name: workspace_id
          in: query
          required: false
          schema:
            anyOf:
              - type: string
              - type: 'null'
            title: Workspace Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProviderKeyListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
        - provider-keys
      summary: Create Provider Key
      operationId: create_provider_key_api_v1_provider_keys_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProviderKeyCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProviderKeyResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    ChatMessage:
      properties:
        workspaceId:
          type: string
          title: Workspaceid
        message:
          type: string
          title: Message
        agentName:
          anyOf:
            - type: string
            - type: 'null'
          title: Agentname
      type: object
      required:
        - workspaceId
        - message
      title: ChatMessage
...
    CreateWorkspaceRequest:
      properties:
        name:
          anyOf:
            - type: string
            - type: 'null'
          title: Name
      type: object
      title: CreateWorkspaceRequest
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    MemoryCreate:
      properties:
        type:
          type: string
          maxLength: 100
          minLength: 1
          title: Type
        domain:
          anyOf:
            - type: string
              maxLength: 100
            - type: 'null'
          title: Domain
        title:
          anyOf:
            - type: string
            - type: 'null'
          title: Title
        summary:
          anyOf:
            - type: string
            - type: 'null'
          title: Summary
        content:
          anyOf:
            - type: string
            - type: 'null'
          title: Content
        metadata:
          anyOf:
            - additionalProperties: true
              type: object
            - type: 'null'
          title: Metadata
        tags:
          anyOf:
            - items:
                type: string
              type: array
            - type: 'null'
          title: Tags
        workspace_id:
          anyOf:
            - type: string
            - type: 'null'
          title: Workspace Id
        source_type:
          anyOf:
            - type: string
            - type: 'null'
          title: Source Type
        source_uri:
          anyOf:
            - type: string
            - type: 'null'
          title: Source Uri
        source_label:
          anyOf:
            - type: string
            - type: 'null'
          title: Source Label
        connector_id:
          anyOf:
            - type: string
            - type: 'null'
          title: Connector Id
        supersedes_id:
          anyOf:
            - type: string
              format: uuid
            - type: 'null'
          title: Supersedes Id
      type: object
      required:
        - type
      title: MemoryCreate
...
    MemorySearch:
      properties:
        query:
          type: string
          minLength: 1
          title: Query
        type:
          anyOf:
            - type: string
            - type: 'null'
          title: Type
        domain:
          anyOf:
            - type: string
            - type: 'null'
          title: Domain
        tags:
          anyOf:
            - items:
                type: string
              type: array
            - type: 'null'
          title: Tags
        top_k:
          type: integer
          maximum: 100.0
          minimum: 1.0
          title: Top K
          default: 10
        threshold:
          anyOf:
            - type: number
              maximum: 1.0
              minimum: 0.0
            - type: 'null'
          title: Threshold
          default: 0.7
      type: object
      required:
        - query
      title: MemorySearch
    MemorySearchResult:
      properties:
        memory:
          $ref: '#/components/schemas/MemoryResponse'
        score:
          type: number
          title: Score
      type: object
      required:
        - memory
        - score
      title: MemorySearchResult
...
    WorkspaceResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        user_id:
          type: string
          format: uuid
          title: User Id
        name:
          type: string
          title: Name
        description:
          anyOf:
            - type: string
            - type: 'null'
          title: Description
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
        - id
        - user_id
        - name
        - created_at
        - updated_at
      title: WorkspaceResponse